// LocalGlovoAngularClient/src/app/models/add-comercio.ts
// Generated by https://quicktype.io

export interface AddComercio {
    id:            string;
    imagen:        string;
    name:          string;
    latitud:       number;
    longitud:      number;
    rating:        number;
    nameDirection: string;
    categorias:    string[];
    productos:     null;
    favoritoList:  null;
}